Added Polish and Italian languages

+++ tools/i18n.py
@@ -17,12 +17,18 @@ for dirname, dirnames, filenames in os.walk('../OpenNumismat'):
         if fileExtension == '.py':
             srcFiles.append(os.path.join(dirname, filename))
 
-for locale in ['ru', 'uk', 'es', 'hu', 'pt', 'el', 'de', 'cs']:
-    outputfile = 'lang_%s.ts' % locale
+f = open('langs')
+langs = [x.strip('\n') for x in f.readlines()]
+
+for lang in langs:
+    if lang == 'en':
+        continue
+    outputfile = 'lang_%s.ts' % lang
     os.system(' '.join([lupdatePath, ' '.join(srcFiles), '-ts', outputfile]))
     os.system(' '.join([linguistPath, outputfile]))
-    dst_file = '../OpenNumismat/lang_%s.qm' % locale
+    dst_file = '../OpenNumismat/lang_%s.qm' % lang
     os.system(' '.join([lreleasePath, outputfile, '-qm', dst_file]))
 
-    src_file = os.path.join(translationsPath, "qt_%s.qm" % locale)
-    shutil.copy(src_file, "../OpenNumismat")
+    src_file = os.path.join(translationsPath, "qtbase_%s.qm" % lang)
+    if os.path.isfile(src_file):
+        shutil.copy(src_file, "../OpenNumismat")
